+++ src/llatvfl/utils/metric.h
@@ -9,9 +9,9 @@
 #include <array>
 using namespace std;
 
-double trapz(vector<double> x, vector<double> y)
+float trapz(vector<float> x, vector<float> y)
 {
-    double res = 0;
+    float res = 0;
     int num_elements = x.size();
     for (int i = 1; i < num_elements; i++)
     {
@@ -20,10 +20,10 @@ double trapz(vector<double> x, vector<double> y)
     return res;
 }
 
-vector<double> get_thresholds_idxs(vector<double> y_pred)
+vector<float> get_thresholds_idxs(vector<float> y_pred)
 {
-    vector<double> thresholds_idxs;
-    set<double> s{};
+    vector<float> thresholds_idxs;
+    set<float> s{};
     reverse(y_pred.begin(), y_pred.end());
     int y_pred_size = y_pred.size();
     for (int i = 0; i < y_pred_size; i++)
@@ -36,7 +36,7 @@ vector<double> get_thresholds_idxs(vector<double> y_pred)
     return thresholds_idxs;
 }
 
-double roc_auc_score(vector<double> y_pred, vector<int> y_true)
+float roc_auc_score(vector<float> y_pred, vector<int> y_true)
 {
     int num_elements = y_pred.size();
     vector<int> temp_idxs(num_elements);
@@ -51,9 +51,9 @@ double roc_auc_score(vector<double> y_pred, vector<int> y_true)
     }
     sort(y_pred.begin(), y_pred.end());
 
-    vector<double> thresholds_idxs = get_thresholds_idxs(y_pred);
+    vector<float> thresholds_idxs = get_thresholds_idxs(y_pred);
 
-    vector<double> tps = {0};
+    vector<float> tps = {0};
     for (int i = 1; i < num_elements; i++)
     {
         tps.push_back(y_true[i] + tps[i - 1]);
@@ -63,7 +63,7 @@ double roc_auc_score(vector<double> y_pred, vector<int> y_true)
         tps[i] = tps[i] / tps[tps.size() - 1];
     }
 
-    vector<double> fps = {1};
+    vector<float> fps = {1};
     for (int i = 1; i < num_elements; i++)
     {
         fps.push_back(1 - y_true[i] + fps[i - 1]);
