+++ src/llatvfl/xgboost/xgboost.h
@@ -10,33 +10,33 @@ using namespace std;
 
 struct XGBoostBase : TreeModelBase<XGBoostParty>
 {
-    double subsample_cols;
-    double min_child_weight;
+    float subsample_cols;
+    float min_child_weight;
     int depth;
     int min_leaf;
-    double learning_rate;
+    float learning_rate;
     int boosting_rounds;
-    double lam;
-    double gamma;
-    double eps;
+    float lam;
+    float gamma;
+    float eps;
     bool use_ispure;
     int active_party_id;
     int completelly_secure_round;
-    double init_value;
+    float init_value;
     int n_job;
     bool save_loss;
 
-    vector<double> init_pred;
+    vector<float> init_pred;
     vector<XGBoostTree> estimators;
-    vector<double> logging_loss;
+    vector<float> logging_loss;
 
-    XGBoostBase(double subsample_cols_ = 0.8,
-                double min_child_weight_ = -1 * numeric_limits<double>::infinity(),
+    XGBoostBase(float subsample_cols_ = 0.8,
+                float min_child_weight_ = -1 * numeric_limits<float>::infinity(),
                 int depth_ = 5, int min_leaf_ = 5,
-                double learning_rate_ = 0.4, int boosting_rounds_ = 5,
-                double lam_ = 1.5, double gamma_ = 1, double eps_ = 0.1,
+                float learning_rate_ = 0.4, int boosting_rounds_ = 5,
+                float lam_ = 1.5, float gamma_ = 1, float eps_ = 0.1,
                 int active_party_id_ = -1, int completelly_secure_round_ = 0,
-                double init_value_ = 1.0, int n_job_ = 1, bool save_loss_ = true)
+                float init_value_ = 1.0, int n_job_ = 1, bool save_loss_ = true)
     {
         subsample_cols = subsample_cols_;
         min_child_weight = min_child_weight_;
@@ -54,10 +54,10 @@ struct XGBoostBase : TreeModelBase<XGBoostParty>
         save_loss = save_loss_;
     }
 
-    virtual vector<double> get_grad(vector<double> &y_pred, vector<double> &y) = 0;
-    virtual vector<double> get_hess(vector<double> &y_pred, vector<double> &y) = 0;
-    virtual double get_loss(vector<double> &y_pred, vector<double> &y) = 0;
-    virtual vector<double> get_init_pred(vector<double> &y) = 0;
+    virtual vector<float> get_grad(vector<float> &y_pred, vector<float> &y) = 0;
+    virtual vector<float> get_hess(vector<float> &y_pred, vector<float> &y) = 0;
+    virtual float get_loss(vector<float> &y_pred, vector<float> &y) = 0;
+    virtual vector<float> get_init_pred(vector<float> &y) = 0;
 
     void load_estimators(vector<XGBoostTree> _estimators)
     {
@@ -69,10 +69,10 @@ struct XGBoostBase : TreeModelBase<XGBoostParty>
         return estimators;
     }
 
-    void fit(vector<XGBoostParty> &parties, vector<double> &y)
+    void fit(vector<XGBoostParty> &parties, vector<float> &y)
     {
         int row_count = y.size();
-        vector<double> base_pred;
+        vector<float> base_pred;
         if (estimators.size() == 0)
         {
             init_pred = get_init_pred(y);
@@ -86,7 +86,7 @@ struct XGBoostBase : TreeModelBase<XGBoostParty>
 
             for (int i = 0; i < estimators.size(); i++)
             {
-                vector<double> pred_temp = estimators[i].get_train_prediction();
+                vector<float> pred_temp = estimators[i].get_train_prediction();
                 for (int j = 0; j < row_count; j++)
                     base_pred[j] += learning_rate * pred_temp[j];
             }
@@ -94,13 +94,13 @@ struct XGBoostBase : TreeModelBase<XGBoostParty>
 
         for (int i = 0; i < boosting_rounds; i++)
         {
-            vector<double> grad = get_grad(base_pred, y);
-            vector<double> hess = get_hess(base_pred, y);
+            vector<float> grad = get_grad(base_pred, y);
+            vector<float> hess = get_hess(base_pred, y);
 
             XGBoostTree boosting_tree = XGBoostTree();
             boosting_tree.fit(&parties, y, grad, hess, min_child_weight,
                               lam, gamma, eps, min_leaf, depth, active_party_id, (completelly_secure_round > i), n_job);
-            vector<double> pred_temp = boosting_tree.get_train_prediction();
+            vector<float> pred_temp = boosting_tree.get_train_prediction();
             for (int j = 0; j < row_count; j++)
                 base_pred[j] += learning_rate * pred_temp[j];
 
@@ -113,15 +113,15 @@ struct XGBoostBase : TreeModelBase<XGBoostParty>
         }
     }
 
-    vector<double> predict_raw(vector<vector<double>> &X)
+    vector<float> predict_raw(vector<vector<float>> &X)
     {
         int row_count = X.size();
-        vector<double> y_pred;
+        vector<float> y_pred;
         copy(init_pred.begin(), init_pred.end(), back_inserter(y_pred));
         int estimators_num = estimators.size();
         for (int i = 0; i < estimators_num; i++)
         {
-            vector<double> y_pred_temp = estimators[i].predict(X);
+            vector<float> y_pred_temp = estimators[i].predict(X);
             for (int j = 0; j < row_count; j++)
                 y_pred[j] += learning_rate * y_pred_temp[j];
         }
@@ -134,10 +134,10 @@ struct XGBoostClassifier : public XGBoostBase
 {
     using XGBoostBase::XGBoostBase;
 
-    double get_loss(vector<double> &y_pred, vector<double> &y)
+    float get_loss(vector<float> &y_pred, vector<float> &y)
     {
-        double loss = 0;
-        double n = y_pred.size();
+        float loss = 0;
+        float n = y_pred.size();
         for (int i = 0; i < n; i++)
         {
             if (y[i] == 1)
@@ -152,38 +152,38 @@ struct XGBoostClassifier : public XGBoostBase
         return loss;
     }
 
-    vector<double> get_grad(vector<double> &y_pred, vector<double> &y)
+    vector<float> get_grad(vector<float> &y_pred, vector<float> &y)
     {
         int element_num = y_pred.size();
-        vector<double> grad(element_num);
+        vector<float> grad(element_num);
         for (int i = 0; i < element_num; i++)
             grad[i] = sigmoid(y_pred[i]) - y[i];
         return grad;
     }
 
-    vector<double> get_hess(vector<double> &y_pred, vector<double> &y)
+    vector<float> get_hess(vector<float> &y_pred, vector<float> &y)
     {
         int element_num = y_pred.size();
-        vector<double> hess(element_num);
+        vector<float> hess(element_num);
         for (int i = 0; i < element_num; i++)
         {
-            double temp_proba = sigmoid(y_pred[i]);
+            float temp_proba = sigmoid(y_pred[i]);
             hess[i] = temp_proba * (1 - temp_proba);
         }
         return hess;
     }
 
-    vector<double> get_init_pred(vector<double> &y)
+    vector<float> get_init_pred(vector<float> &y)
     {
-        vector<double> init_pred(y.size(), init_value);
+        vector<float> init_pred(y.size(), init_value);
         return init_pred;
     }
 
-    vector<double> predict_proba(vector<vector<double>> &x)
+    vector<float> predict_proba(vector<vector<float>> &x)
     {
-        vector<double> raw_score = predict_raw(x);
+        vector<float> raw_score = predict_raw(x);
         int row_count = x.size();
-        vector<double> predicted_probas(row_count);
+        vector<float> predicted_probas(row_count);
         for (int i = 0; i < row_count; i++)
             predicted_probas[i] = sigmoid(raw_score[i]);
         return predicted_probas;
